+++ tests/testthat/test-syntax_highlight.R
@@ -1,0 +1,7 @@
+context("Checking syntax_highlight")
+
+test_that("syntax_highlight ...",{
+
+
+})
+